edited price to be integer instead of decimal

+++ backend/db/migrations/20220627181005-create-spot.js
@@ -39,7 +39,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.DECIMAL(8, 2)
+        type: Sequelize.INTEGER(6)
       },
       createdAt: {
         allowNull: false,
